+++ crates/tauri-store-macros/src/collection.rs
@@ -43,91 +43,91 @@ pub fn impl_collection(ast: &DeriveInput) -> TokenStream {
           self.0.with_store(id, f)
         }
 
-        /// Gets a clone of the store state if it exists.
-        pub fn state(&self, store_id: impl AsRef<str>) -> Result<StoreState> {
-          self.0.state(store_id)
+        /// Gets a clone of the raw store state if it exists.
+        pub fn raw_state(&self, store_id: impl AsRef<str>) -> Result<StoreState> {
+          self.0.raw_state(store_id)
         }
 
         /// Gets the store state, then tries to parse it as an instance of type `T`.
-        pub fn try_state<T>(&self, store_id: impl AsRef<str>) -> Result<T>
+        pub fn state<T>(&self, store_id: impl AsRef<str>) -> Result<T>
         where
           T: DeserializeOwned,
         {
-          self.0.try_state(store_id)
+          self.0.state(store_id)
         }
 
         /// Gets the store state, then tries to parse it as an instance of type `T`.
         ///
         /// If it cannot be parsed, returns the provided default value.
-        pub fn try_state_or<T>(&self, store_id: impl AsRef<str>, default: T) -> Result<T>
+        pub fn state_or<T>(&self, store_id: impl AsRef<str>, default: T) -> Result<T>
         where
           T: DeserializeOwned,
         {
-          self.0.try_state_or(store_id, default)
+          self.0.state_or(store_id, default)
         }
 
         /// Gets the store state, then tries to parse it as an instance of type `T`.
         ///
         /// If it cannot be parsed, returns the default value of `T`.
-        pub fn try_state_or_default<T>(&self, store_id: impl AsRef<str>) -> Result<T>
+        pub fn state_or_default<T>(&self, store_id: impl AsRef<str>) -> Result<T>
         where
           T: DeserializeOwned + Default,
         {
-          self.0.try_state_or_default(store_id)
+          self.0.state_or_default(store_id)
         }
 
         /// Gets the store state, then tries to parse it as an instance of type `T`.
         ///
         /// If it cannot be parsed, returns the result of the provided closure.
-        pub fn try_state_or_else<T>(&self, store_id: impl AsRef<str>, f: impl FnOnce() -> T) -> Result<T>
+        pub fn state_or_else<T>(&self, store_id: impl AsRef<str>, f: impl FnOnce() -> T) -> Result<T>
         where
           T: DeserializeOwned,
         {
-          self.0.try_state_or_else(store_id, f)
+          self.0.state_or_else(store_id, f)
         }
 
-        /// Gets a value from a store.
-        pub fn get(&self, store_id: impl AsRef<str>, key: impl AsRef<str>) -> Option<Json> {
-          self.0.get(store_id, key)
+        /// Gets a raw value from a store.
+        pub fn get_raw(&self, store_id: impl AsRef<str>, key: impl AsRef<str>) -> Option<Json> {
+          self.0.get_raw(store_id, key)
         }
 
         /// Gets a value from a store and tries to parse it as an instance of type `T`.
-        pub fn try_get<T>(&self, store_id: impl AsRef<str>, key: impl AsRef<str>) -> Result<T>
+        pub fn get<T>(&self, store_id: impl AsRef<str>, key: impl AsRef<str>) -> Result<T>
         where
           T: DeserializeOwned,
         {
-          self.0.try_get(store_id, key)
+          self.0.get(store_id, key)
         }
 
         /// Gets a value from a store and tries to parse it as an instance of type `T`.
         ///
         /// If the key does not exist, returns the provided default value.
-        pub fn try_get_or<T>(&self, store_id: impl AsRef<str>, key: impl AsRef<str>, default: T) -> T
+        pub fn get_or<T>(&self, store_id: impl AsRef<str>, key: impl AsRef<str>, default: T) -> T
         where
           T: DeserializeOwned,
         {
-          self.0.try_get_or(store_id, key, default)
+          self.0.get_or(store_id, key, default)
         }
 
         /// Gets a value from a store and tries to parse it as an instance of type `T`.
         ///
         /// If the key does not exist, returns the default value of `T`.
-        pub fn try_get_or_default<T>(&self, store_id: impl AsRef<str>, key: impl AsRef<str>) -> T
+        pub fn get_or_default<T>(&self, store_id: impl AsRef<str>, key: impl AsRef<str>) -> T
         where
           T: Default + DeserializeOwned,
         {
-          self.0.try_get_or_default(store_id, key)
+          self.0.get_or_default(store_id, key)
         }
 
         /// Gets a value from a store and tries to parse it as an instance of type `T`.
         ///
         /// If the key does not exist, returns the result of the provided closure.
-        pub fn try_get_or_else<T, F>(&self, store_id: impl AsRef<str>, key: impl AsRef<str>, f: F) -> T
+        pub fn get_or_else<T, F>(&self, store_id: impl AsRef<str>, key: impl AsRef<str>, f: F) -> T
         where
           T: DeserializeOwned,
           F: FnOnce() -> T,
         {
-          self.0.try_get_or_else(store_id, key, f)
+          self.0.get_or_else(store_id, key, f)
         }
 
         /// Sets a key-value pair in a store.