+++ examples/pinia-nuxt/nuxt.config.ts
@@ -1,5 +1,5 @@
 export default defineNuxtConfig({
-  compatibilityDate: '2024-11-01',
+  compatibilityDate: 'latest',
   modules: ['@pinia/nuxt'],
   ssr: false,
   devServer: {