+++ examples/svelte/src/routes/+page.svelte
@@ -1,7 +1,6 @@
 <script lang="ts">
   import { openStore, store } from '$lib/store';
   import { exit } from '@tauri-apps/plugin-process';
-  import { printStore } from '$lib/commands';
   import { clearAutosave, setAutosave } from '@tauri-store/svelte/src/lib/index.js';
 
   function onKeyDown(event: KeyboardEvent) {
@@ -22,17 +21,14 @@
   <section id="counter">
     <p>
       Counter: {$store.counter}
-      Nested: {$store.nested.foo.bar.baz}
     </p>
     <div class="action">
       <button type="button" onclick={() => $store.counter++}>Increment</button>
-      <button type="button" onclick={() => $store.nested.foo.bar.baz++}>Increment Nested</button>
       <button type="button" onclick={() => store.start()}>Start</button>
       <button type="button" onclick={() => store.stop()}>Stop</button>
       <button type="button" onclick={() => store.save()}>Save</button>
       <button type="button" onclick={() => store.saveNow()}>Save Now</button>
-      <button type="button" onclick={printStore}>Print</button>
-      <button type="button" onclick={openStore}>Open</button>
+      <button type="button" onclick={() => openStore()}>Open</button>
     </div>
   </section>
 </main>