+++ examples/vue/src/App.vue
@@ -1,5 +1,4 @@
 <script setup lang="ts">
-import { printStore } from './commands';
 import { onKeyDown } from '@vueuse/core';
 import { openStore, store } from './store';
 import { exit } from '@tauri-apps/plugin-process';
@@ -17,15 +16,13 @@ onKeyDown('Escape', () => void exit(0));
       <button type="button" @click="clearAutosave">Clear Autosave</button>
     </div>
     <section id="counter">
-      <p>Counter: {{ store.counter }} Nested: {{ store.nested.foo.bar.baz }}</p>
+      <p>Counter: {{ store.counter }}</p>
       <div class="action">
         <button type="button" @click="() => store.counter++">Increment</button>
-        <button type="button" @click="() => store.nested.foo.bar.baz++">Increment Nested</button>
         <button type="button" @click="start">Start</button>
         <button type="button" @click="stop">Stop</button>
         <button type="button" @click="save">Save</button>
         <button type="button" @click="saveNow">Save Now</button>
-        <button type="button" @click="printStore">Print</button>
         <button type="button" @click="openStore">Open</button>
       </div>
     </section>